requirements.txt
langchain
langchain-community
langchain-groq
langchain-google-genai
google-generativeai
chromadb
python-dotenv
pydantic
streamlit
pypdf
faiss-cpu   
sentence-transformers